MINOR/BUG: fixed memory reallocation function otc_dbg_realloc()
When reallocating memory, no attention was paid to whether the memory
was initially allocated via the memory allocator wrapper (debugger) or
directly using the malloc library.

Another problem (but less important because it did not cause the programs
using this library to crash) was that the otc_dbg_mem_alloc() function
did not properly record memory reallocation in its statistics.