MINOR/BUG: fixed calculation of data size in the otc_text_map structure
When increasing the number of data pairs in the otc_text_map structure,
this value was miscalculated for unallocated data (ie those for which
the number of pairs was 0).

The newly allocated memory was cleared, but this was done at the wrong
offset.

Also, there was an error in the macro OT_TEXT_MAP_SIZE() where sizeof
was taken from the wrong data type.