MINOR: separating the library load from running the tracer thread
Two functions have been added, otc_tracer_load() and otc_tracer_start(),
which are actually the otc_tracer_init() function divided into two parts:
the dynamic loading of a tracer library (a tracer plugin) and the launch
of the tracer thread (i.e. initialization).

This is done so that the dynamic loading of the library can be separated
from the running of the tracer thread.  For example, if we use this in a
daemon program, we can load the library before switching to daemon mode
(before the fork()/exec() functions) and then initialize the tracing
process in the child process.

The otc_tracer_init() function has not been removed from the library,
and can still be used in the previous way.

Also, the library version has been changed to 1.1.0.